trying heuristic with all possible set of destinations, now I will have to hadd a visualization for the tree and check it's validity

+++ src/main/java/ch/usi/inf/dslab/bftamcast/treesearch/Vertex.java
@@ -10,11 +10,11 @@ import java.util.List;
  */
 public class Vertex {
 	public int ID;
-	public int capacity, resCapacity;
+	public double capacity, resCapacity;
 	List<Vertex> connections = new ArrayList<>();
 	Vertex parent;
 
-	public Vertex(int ID, String conf, int capacity) {
+	public Vertex(int ID, String conf, double capacity) {
 		this.ID = ID;
 		this.capacity = capacity;
 		this.resCapacity = capacity;
